Rename async module to asyncparser
