+++ tox.ini
@@ -1,9 +1,8 @@
 [tox]
-envlist =py27,py34,py35,py36
+envlist =py27,py34,py35,py36,py37
 [testenv]
 deps=
-  py26: ordereddict
   -rdev-requirements.txt
-  {py34,py35,py36}: -rdev-requirements-py3.txt
+  {py34,py35,py36,py37}: -rdev-requirements-py3.txt
 commands=
   invoke test
